requirements-pip.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
# autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
# Brotli
cachetools==5.3.1
certifi
cffi
charset-normalizer
comm==0.1.3
contourpy==1.1.0
cryptography
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dotmap==1.3.30
einops==0.6.1
executing==1.2.0
fastjsonschema==2.17.1
filelock
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.6.0
gmpy2
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.16.4
idna
isoduration==20.11.0
jedi==0.18.2
Jinja2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe
mccabe==0.7.0
miniimagenettools==0.2.2
mistune==3.0.1
mpmath
nest-asyncio==1.5.6
networkx
notebook_shim==0.2.3
numpy
oauthlib==3.2.2
opencv-python==4.8.0.74
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser
Pygments==2.15.1
pyOpenSSL
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks
python-box==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pyzmq==25.1.0
requests
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy
terminado==0.17.1
threadpoolctl==3.1.0
timm==0.9.2
tinycss2==1.2.1
tomli
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm 
traitlets==5.9.0
triton==2.0.0
typing_extensions
uri-template==1.3.0
urllib3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
